requirements.txt
annotated-types==0.7.0
asgiref==3.9.2
brotli==1.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cloudinary==1.44.1
colorama==0.4.6
cssselect2==0.8.0
dj-database-url==3.0.1
Django==5.2.6
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
fonttools==4.61.1
gunicorn==23.0.0
h11==0.16.0
humanize==4.13.0
idna==3.10
numpy==2.4.2
openpyxl==3.1.5
packaging==25.0
pandas==3.0.0
pillow==12.0.0
psycopg2-binary==2.9.10
pycparser==3.0
pydantic==2.11.9
pydantic_core==2.33.2
pydyf==0.12.1
PyJWT==2.10.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tinycss2==1.5.1
tinyhtml5==2.0.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
weasyprint==68.0
webencodings==0.5.1
whitenoise==6.11.0
zopfli==0.4.0
